% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{comprueba}
\alias{comprueba}
\title{Comprobaciones para informes y presentaciones}
\usage{
comprueba(file_name = NULL, doc_format = NULL)
}
\arguments{
\item{file_name}{Cadena de caracteres con el nombre del archivo sin
extensión. Por defecto, los archivos recibirán un nombre genérico
\code{informe_form}, donde "form" varía en función del formato del informe:
pdf, html, odt, docx, etc).}

\item{doc_format}{Cadena de caracteres con el formato del documento. Los
formatos admitidos son pdf, latex, html, docx, odt, pres_html y beamer}
}
\description{
Esta función crea una lista con todas las especificaciones necesarias para
crear un informe o presentación siguiendo el formato de la plantilla de
fisabior.
}
