% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{copia_csl}
\alias{copia_csl}
\title{Copia el archivo csl para estilizar la bibliografía y lo introduce en el
directorio donde se encuentra el informe.}
\usage{
copia_csl(path)
}
\arguments{
\item{path}{Ruta al documento.}
}
\description{
Copia el archivo csl para estilizar la bibliografía y lo introduce en el
directorio donde se encuentra el informe.
}
