% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{copia_imagenes}
\alias{copia_imagenes}
\title{Copia imágenes y las introduce en el subdirectorio figuras se encuentra el
informe.}
\usage{
copia_imagenes(path)
}
\arguments{
\item{path}{Ruta al documento.}
}
\description{
Copia imágenes y las introduce en el subdirectorio figuras se encuentra el
informe.
}
