% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{copia_referencias}
\alias{copia_referencias}
\title{Copia el archivo bib con las referencias de prueba y lo introduce en el
directorio donde se encuentra el informe.}
\usage{
copia_referencias(path)
}
\arguments{
\item{path}{Ruta al documento.}
}
\description{
Copia el archivo bib con las referencias de prueba y lo introduce en el
directorio donde se encuentra el informe.
}
