% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisabior.R
\docType{package}
\name{fisabior}
\alias{fisabior}
\alias{fisabior-package}
\title{fisabior}
\description{
Librería con funciones, datos y utilidades de uso habitual para los miembros
del grupo de investigación del Área de Desigualdades de la Fundación para el
Fomento de la Investigación Sanitaria y Biomédica de la Comunidad Valenciana
(FISABIO)
}

