% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{informe}
\alias{informe}
\title{Funci贸n que devuelve el formato de salida apropiado para generar el informe o
la presentaci贸n en PDF.}
\usage{
informe(beamer = FALSE, ...)
}
\arguments{
\item{...}{Otros argumentos a pasar.}
}
\value{
Objeto con clase "rmarkdown_output_format".
}
\description{
Esta funci贸n crea una lista con todas las especificaciones necesarias para
crear un PDF siguiendo el formato de la plantilla de fisabior. Por defecto se
usa XeLaTeX como motor de LaTeX.
}
\details{
La funci贸n solo debe utilizarse desde una llamada a
  rmarkdown::render() o dentro de un documento .Rmd
}
