% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{informe_pdf}
\alias{informe_pdf}
\title{Función que devuelve el formato de salida apropiado para generar el informe en PDF.}
\usage{
informe_pdf(keep_tex = TRUE, md_extensions = "-autolink_bare_uris")
}
\arguments{
\item{keep_tex}{Lógico: ¿debe guardarse el archivo .tex?}

\item{md_extensions}{Cadena de caracteres indicando extensiones a markdown desde pandoc.}
}
\value{
Objeto con clase 'rmarkdown_output_format'.
}
\description{
Función que devuelve el formato de salida apropiado para generar el informe en PDF.
}
\examples{
informe_pdf(keep_tex = TRUE,
            md_extensions = '-autolink_bare_uris')
}

