% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{presentacion_beamer}
\alias{presentacion_beamer}
\title{Crea una presentación estadística en PDF -- LaTeX -- Beamer}
\usage{
presentacion_beamer(file_name = "presentacion")
}
\arguments{
\item{file_name}{Cadena de caracteres con el nombre del archivo sin
extensión. Por defecto, los archivos recibirán un nombre genérico
\code{informe}.}
}
\value{
Crea un documento principal y un conjunto de documentos de respaldo
  (algunos directamente en el directorio de presentacion/latex y otros en los
  directorios de data/cache o figuras/presentacion).
}
\description{
Crea automáticamente una presentación estadística en PDF -- LaTeX -- Beamer
empleando una plantilla base con detalles del grupo. El documento emplea
formato LaTeX puro y tiene extensión .Rnw (no utiliza en ningún caso
Markdown).
}
\details{
La función solo funciona si el directorio de trabajo se circunscribe
  a un proyecto de RStudio, es decir, si existe un archivo \emph{*.Rproj} en
  el directorio de trabajo. Si el proyecto se ha creado con
  \code{\link{init_proj}}, esto no debería causar molestias, pues se crea un
  \emph{*.Rproj} por defecto.

  Es necesario tener instalada una distribución LaTeX que incorpore (como
  mínimo) los siguientes paquetes: \code{adjustbox}, \code{amsmath},
  \code{amssymb}, \code{array}, \code{authblk}, \code{beamer},
  \code{biblatex}, \code{booktabs}, \code{caption}, \code{ccicons},
  \code{csquotes}, \code{float}, \code{fontenc}, \code{fontspec},
  \code{footmisc}, \code{graphicx}, \code{hyperref}, \code{letltxmacro},
  \code{longtable}, \code{lscape}, \code{mathtools}, \code{microtype},
  \code{multirow}, \code{parskip}, \code{polyglossia}, \code{rotating},
  \code{setspace}, \code{tabularx}, \code{textpos}, \code{tikz},
  \code{xltxtra}, \code{xunicode} y el tema \code{motropolis}.

  El proceso genera un documento rmarkdown que produce un PDF tras ser
  compilado con knitr y XeLaTeX. La bibliografía se gestiona a través de
  \code{biblatex} empleando el gestor de referencias \code{biber}, de modo
  que este último debe estar instalado en el sistema.

  Con todo, en primer lugar la función comprueba que el proyecto tenga
  declaradas todas estas opciones (vienen por defecto al emplear
  \code{\link{init_proj}}) y que el software apropiado esté instalado. Si
  no hubiera un acuerdo entre lo esperado y lo declarado en las opciones del
  proyecto, la función pregunta qué hacer para que, o bien se cambien las
  opciones del proyecto de forma automática o se detenga la ejecución y se
  cambien a mano.
}
\examples{
\dontrun{
library(fisabior)
init_proj(proj_nom = "proyecto_europeo_X",
          proj_dir = "~/proyectos",
          git      = TRUE)

presentacion_beamer(file_name  = "presentacion_proyecto_x")
}
}
