% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{informe}
\alias{informe}
\title{Crea un informe estadístico desde una plantilla}
\usage{
informe(formato = "pdf", titulo_proj = NULL, titulo_doc = NULL)
}
\arguments{
\item{formato}{Cadena de caracteres con el formato deseado para el informe.
Las opciones admitidas son pdf, html, docx y odt. Por defecto se selecciona
pdf.}

\item{titulo_proj}{Cadena de caracteres con el título del informe. Por
defecto se asigna 'Análisis del proyecto X', donde X es el nombre del
directorio principal del proyecto.}

\item{titulo_doc}{Cadena de caracteres con el nombre del archivo sin
extensión. Por defecto se asigna el nombre 'informe_form.ext', donde 'form'
(formato) y 'ext' (extensión) varían en función de lo seleccionado en en el
argumento formato.}
}
\value{
En función del formato escogido, se crea un documento principal y un
  conjunto de documentos de respaldo. El comportamiento inicial es abrir el
  documento principal con el editor.
}
\description{
Crea automáticamente un informe estadístico empleando una plantilla base con
detalles del grupo, ofreciendo la posibilidad de que el documento generado
tenga diversos formatos (PDF, HTML, DOCX u ODT).
}
\details{
La función solo funciona si en el directorio de trabajo existe un
  proyecto de RStudio, es decir, un archivo '.Rproj': si el proyecto se ha
  creado con fisabior::init_proj(), esto no debería causar molestias, pues se
  crea un .Rproj por defecto.

  Esta función requiere tener operativo los paquetes knitr y rmarkdown. Si se
  desea generar informes PDF, también es necesario tener una distribución
  LaTeX que incorpore (como mínimo) los siguientes paquetes: adjustbox,
  amsmath, array, authblk, biblatex, booktabs, caption, csquotes, float,
  fontenc, fontspec, footmisc, graphicx, hyperref, longtable, lscape,
  mathtools, memoir, multirow, polyglossia, rotating, setspace, tikz,
  tkz-graph, xltxtra, xunicode. La opción por defecto es un documento LaTeX
  compilado con XeLaTeX (en lugar de pdfLaTeX, ya que el primero ofrece una
  mayor compatibilidad de fuentes). Con todo, la función comprueba en primer
  lugar que el proyecto tenga declaradas estas opciones (vienen por defecto
  al emplear init_proj). Si no hubiera un acuerdo entre lo esperado y lo
  declarado en las opciones del proyecto, la función pregunta qué hacer para
  que, o bien se cambien las opciones del proyecto de forma automática o se
  detenga la ejecución y se cambien a mano.
}
\examples{
informe(formato     = 'pdf',
        titulo_proj = 'Informe Estadístico de Prueba',
        titulo_doc  = 'informe_prueba')
}

