% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informes.R
\name{informe_pdf}
\alias{informe_pdf}
\title{Crea un informe estadístico en PDF -- Markdown}
\usage{
informe_pdf(file_name = "informe")
}
\arguments{
\item{file_name}{Cadena de caracteres con el nombre del archivo sin
extensión. Por defecto, los archivos recibirán un nombre genérico
\code{informe}.}
}
\value{
Crea un documento principal y un conjunto de documentos de respaldo
  (algunos directamente en el directorio de informes/pdf y otros en los
  directorios de data/cache o figuras/pdf).
}
\description{
Crea automáticamente un informe estadístico en PDF -- Markdown empleando una
plantilla base con detalles del grupo.
}
\details{
La función solo funciona si el directorio de trabajo se circunscribe
  a un proyecto de RStudio, es decir, si existe un archivo \emph{*.Rproj} en
  el directorio de trabajo. Si el proyecto se ha creado con
  \code{\link{init_proj}}, esto no debería causar molestias, pues se crea
  un \emph{*.Rproj} por defecto.

  Es necesario tener instalada una distribución LaTeX que incorpore (como
  mínimo) los siguientes paquetes: \code{adjustbox}, \code{amsmath},
  \code{amssymb}, \code{array}, \code{authblk}, \code{booktabs},
  \code{caption}, \code{ccicons}, \code{csquotes}, \code{float},
  \code{fontenc}, \code{fontspec}, \code{footmisc}, \code{graphicx},
  \code{hyperref}, \code{letltxmacro}, \code{longtable}, \code{lscape},
  \code{mathtools}, \code{microtype}, \code{multirow}, \code{parskip},
  \code{polyglossia}, \code{rotating}, \code{setspace}, \code{tabularx},
  \code{textpos}, \code{tikz}, \code{xltxtra}, \code{xunicode}.

  El proceso genera un documento R Markdown que produce un PDF tras ser
  compilado con knitr, pandoc y XeLaTeX (en lugar de pdfLaTeX, ya que el
  primero ofrece una mayor compatibilidad de fuentes para idiomas con
  caracteres especiales --caracteres no ASCII, como el nuestro--). La
  bibliografía se gestiona a través de pandoc-citeproc y un archivo csl con
  el estilo (por defecto, estilo APA). Si quieres cambiar el estilo las
  referencias, elimina el archivo csl que se genera por defecto y añade en su
  lugar el que tu desees, cambiando la opción en el YAML del documento .Rmd.

  Con todo, en primer lugar la función comprueba que el proyecto tenga
  declaradas todas estas opciones (vienen por defecto al emplear
  \code{\link{init_proj}}) y que el software apropiado esté instalado. Si
  no hubiera un acuerdo entre lo esperado y lo declarado en las opciones del
  proyecto, la función pregunta qué hacer para que, o bien se cambien las
  opciones del proyecto de forma automática o se detenga la ejecución y se
  cambien a mano.
}
\examples{
\dontrun{
library(fisabior)
init_proj(proj_nom = "proyecto_europeo_X",
          proj_dir = "~/proyectos",
          git      = TRUE)

informe_pdf(file_name = "informe_proyecto_x")
}
}
