% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_proj.R
\name{init_proj}
\alias{init_proj}
\title{Crea un proyecto fisabior}
\usage{
init_proj(proj_nom = NULL, proj_dir = NULL, git = TRUE)
}
\arguments{
\item{proj_nom}{Cadena de caracteres con el nombre a asignar al proyecto. Por
defecto se asigna el nombre "proyecto-yyy-mm-dd".}

\item{proj_dir}{Cadena de caracteres indicando la ruta al directorio
principal donde se creará el proyecto. Por defecto se escoge el directorio
actual de trabajo, devuelto por la función getwd().}

\item{git}{Valor lógico (TRUE o FALSE), indicando si se desea generar un
repositorio git asociado al proyecto.}
}
\value{
La función crea la siguiente estructura de directorios en
  proj_dir/proj_nom/:
\itemize{
  \item datos/
    \itemize{
      \item brutos/
      \item procesados/
      \item cartografia/
  }
  \item figuras/
  \item informes/
  \itemize{
    \item docx/
    \item pdf-markdown/
    \item pdf-latex/
    \item html/
    \item odt/
    \item beamer/
  }
  \item cache/
  \item src/
  \itemize{
    \item bugs/
    \item cpp/
    \item jags/
    \item stan/
  }
  \item configuracion/
  \item r/
  \item articulo/
  \itemize{
    \item enviado/
    \item revision/
    \item proof/
  }
}
}
\description{
Crea automáticamente un proyecto siguiendo la plantilla fisabior, empleando
una jerarquía de directorios unificada, cargando archivos y funciones comunes
del grupo de colaboradores. Opcionalmente, genera un repositorio git.
}
\details{
La función crea la estructura de directorios a partir del nombre del
  proyecto empleando minúsculas, con independencia de si el usuario ha
  introducido el argumento con mayúsculas: P. ej., proj_name = "Proyecto_X"
  siempre será reconocido por la función como proj_name = "proyecto_x".
}
\examples{
\dontrun{
library(fisabior)
init_proj(proj_nom = "proyecto_europeo_X", proj_dir = "~/proyectos")
}
}
